refactor: 디렉토리구조 변경 및 매직넘버 삭제
