+++ src/main/java/domain/quiztype/engnamequiz/objective/subjective/AnagramQuiz.java
@@ -1,23 +1,24 @@
-package domain.quiztype.subjective;
+package domain.quiztype.engnamequiz.objective.subjective;
 
 import domain.Student;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import static domain.QuizConstants.*;
 
 public class AnagramQuiz extends SubjectiveQuiz {
 
-    private static final int ANAGRAM_SCORE = 7;
-
     public AnagramQuiz(Student student) {
         super(student);
+        generateQuiz();
     }
 
+    @Override
     protected void generateQuiz() {
         List<Character> anagram = generateAnagram();
         question = anagram.toString() + "의 알파벳을 가진 학생의 영어이름은?";
-        score = ANAGRAM_SCORE;
+        answer = student.getEngFirstName();
+        score = ANAGRAM_QUIZ_SCORE;
     }
 
     private List<Character> generateAnagram() {
@@ -29,6 +30,4 @@ public class AnagramQuiz extends SubjectiveQuiz {
         Collections.shuffle(ans);
         return ans;
     }
-
-
 }
